moved infogain and related functions to infogain.py
